+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 ## [0.0.1] - 01-24-2021.
 
 ### Added
+
 - HyperDeck Connection Module
+
+## [0.0.2] - 03-28-2021.
+
+### Added
+
+- Hyperdeck Operation Module
+- Hyperdeck Device and transport info